% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail.r
\name{avail.species}
\alias{avail.species}
\title{species names in pecpro database}
\usage{
avail.species(database = NULL)
}
\arguments{
\item{database}{databse in which to look (e.g., pec_pro)}
}
\description{
species names in pecpro database
}
\details{
lists species names in pec_pro. Not all species have all tables available (bio vs commercial samples)
}
