% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{bridge.odbc}
\alias{bridge.odbc}
\title{function to extract data fom odbc databases}
\usage{
bridge.odbc(data, database, species, user, password)
}
\arguments{
\item{data}{data to extract (see avail.data function)}

\item{database}{name of the database (see avail.database function)}

\item{species}{name of the species (see avail.species function)}

\item{user}{user name to connect to database}

\item{password}{password to connect to database}
}
\description{
function to extract data fom odbc databases
}
\details{
workhorse function that will do the extracting
}
