% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.bio}
\alias{get.bio}
\title{function to get (car)bio data fom odbc databases}
\usage{
get.bio(species, user, password)
}
\arguments{
\item{species}{name of the species (see avail.species function)}

\item{user}{user name to connect to database}

\item{password}{password to connect to database}
}
\description{
function to get (car)bio data fom odbc databases
}
\details{
There are small differences with the SAP extractions (e.g., if sample.id=NA rows are not kept, trivial differences in rounding for length/weight)
}
