% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.eggs}
\alias{get.eggs}
\title{function to get egg survey data}
\usage{
get.eggs(user, password, driver = "Oracle in OraClient12Home1")
}
\arguments{
\item{user}{user name to connect to database}

\item{password}{password to connect to database}

\item{driver}{character string (default: Oracle in OraClient12Home1)}
}
\description{
function to get egg survey data
}
\details{
The default driver is 'Oracle in OraClient12Home1'. Use argument driver=xxxx to connect with another driver.
}
