% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.lf}
\alias{get.lf}
\title{function to get length-frequency data fom odbc databases}
\usage{
get.lf(species, user, password)
}
\arguments{
\item{species}{name of the species (see avail.species function)}

\item{user}{user name to connect to database}

\item{password}{password to connect to database}
}
\description{
function to get length-frequency data fom odbc databases
}
\details{
no details
}
