% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{repl}
\alias{repl}
\title{replace levels}
\usage{
repl(x, a, b)
}
\arguments{
\item{x}{string with values to replace}

\item{a}{string with unique old values}

\item{b}{string with unique new levels}
}
\description{
replace levels
}
\details{
Used to replace cryptic numbers in data by meaningful character strings.
}
