% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql.from}
\alias{sql.from}
\title{create string ith table to select from}
\usage{
sql.from(table, species)
}
\arguments{
\item{table}{full table names (without species)}

\item{species}{species name (see avail.species function)}

\item{table_short}{short table names}
}
\description{
create string ith table to select from
}
