% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql.query}
\alias{sql.query}
\title{create sql query}
\usage{
sql.query(table, column, species)
}
\arguments{
\item{table}{full table names (without species)}

\item{column}{column names}

\item{species}{species name (see avail.species function)}

\item{table_short}{short table names}
}
\description{
create sql query
}
