% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql.select}
\alias{sql.select}
\title{create string with columns to be selected}
\usage{
sql.select(table, column)
}
\arguments{
\item{column}{column name}

\item{table_short}{short table name}
}
\description{
create string with columns to be selected
}
