% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{bridge.odbc}
\alias{bridge.odbc}
\title{function to extract data fom odbc databases}
\usage{
bridge.odbc(query, database, user, password)
}
\arguments{
\item{query}{sql query to extract data}

\item{database}{name of the database (see avail.database function)}

\item{user}{user name to connect to database}

\item{password}{password to connect to database}
}
\description{
function to extract data fom odbc databases
}
\details{
connect, extract, close
}
