% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql.where}
\alias{sql.where}
\title{create string for where}
\usage{
sql.where(table, column, arguments)
}
\arguments{
\item{column}{column names}

\item{arguments}{subset for columns}

\item{table_short}{short table names}
}
\description{
create string for where
}
